Final Push, v1.0
Yay!

+++ app/models/votes.js
@@ -3,7 +3,7 @@ module.exports = function(sequelize, DataTypes) {
 	var Vote = sequelize.define('Vote', {
 		ProposalId: 'INT',
 		VoterId: 'MEDIUMINT',
-		Value: 'TINYINT'
+		Value: 'MEDIUMINT'
 	});
 	
 	return Vote;
